requirements.txt
discord.py==2.3.0
openai==1.68.2
beautifulsoup4==4.12.2
youtube_transcript_api==0.6.1
pypdf2==3.0.1
feedparser==6.0.10
trafilatura==1.8.1
groq
anthropic
google-generativeai
replicate==0.32.1
litellm==1.66.3
audioop
